add cluster benchmarks
