+++ workflow/targets/cluster.smk
@@ -12,39 +12,83 @@ CLUSTER_OUTPUTS = {
         CLUSTER_FP
         / "{channel_combo}"
         / "{cell_class}"
-        / "tsvs"
         / get_filename({}, "aggregate_cleaned", "tsv"),
     ],
     "phate_leiden_clustering": [
         CLUSTER_FP
         / "{channel_combo}"
         / "{cell_class}"
-        / "tsvs"
+        / "{leiden_resolution}"
         / get_filename(
-            {"leiden_resolution": "{leiden_resolution}"},
+            {},
             "phate_leiden_clustering",
             "tsv",
         ),
         CLUSTER_FP
         / "{channel_combo}"
         / "{cell_class}"
-        / "plots"
-        / get_filename(
-            {"leiden_resolution": "{leiden_resolution}"}, "cluster_sizes", "png"
-        ),
+        / "{leiden_resolution}"
+        / get_filename({}, "cluster_sizes", "png"),
         CLUSTER_FP
         / "{channel_combo}"
         / "{cell_class}"
-        / "plots"
-        / get_filename({"leiden_resolution": "{leiden_resolution}"}, "clusters", "png"),
+        / "{leiden_resolution}"
+        / get_filename({}, "clusters", "png"),
     ],
     "benchmark_clusters": [
         CLUSTER_FP
         / "{channel_combo}"
         / "{cell_class}"
-        / "plots"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Real"}, "integrated_results", "json"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Shuffled"}, "integrated_results", "json"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Real"}, "combined_table", "tsv"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Shuffled"}, "combined_table", "tsv"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Real"}, "global_metrics", "json"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Shuffled"}, "global_metrics", "json"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Real"}, "pie_chart", "png"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
         / get_filename(
-            {"leiden_resolution": "{leiden_resolution}"}, "cluster_benchmarks", "png"
+            {"cluster_benchmark": "Shuffled"}, "enrichment_pie_chart", "png"
+        ),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename({"cluster_benchmark": "Real"}, "enrichment_bar_chart", "png"),
+        CLUSTER_FP
+        / "{channel_combo}"
+        / "{cell_class}"
+        / "{leiden_resolution}"
+        / get_filename(
+            {"cluster_benchmark": "Shuffled"}, "enrichment_bar_chart", "png"
         ),
     ],
 }
@@ -58,9 +102,9 @@ CLUSTER_OUTPUT_MAPPINGS = {
 
 # TODO: Use all combos
 # cluster_wildcard_combos = cluster_wildcard_combos[
-#     (cluster_wildcard_combos["cell_class"].isin(["mitotic"]))
+#     (cluster_wildcard_combos["cell_class"].isin(["Interphase"]))
 #     & (cluster_wildcard_combos["channel_combo"].isin(["DAPI_COXIV_CENPA_WGA"]))
-#     & (cluster_wildcard_combos["leiden_resolution"].isin([8]))
+#     & (cluster_wildcard_combos["leiden_resolution"].isin([15]))
 # ]
 
 CLUSTER_OUTPUTS_MAPPED = map_outputs(CLUSTER_OUTPUTS, CLUSTER_OUTPUT_MAPPINGS)
